Full rewrite to support easier cross-compiling for package users
Now, the bulk of the logic is implemented in C. The interface to
this implementation is include/go-acl.h, which is intended to be
fully cross-platform (that is, a vanilla cgo installation will
be able to cross-compile for any other architecture without any
extra headers or libraries). This interface is implemented by
C files in src, and these implementations are compiled into .a
files by the package authors; the library files are distributed
with the package itself, and the .go files which use cgo are
written to expect these files to be in place.

The motivation for this change is that go-acl is not a particularly
specialized library, so it's unreasonable to expect that users who
want to both use it and cross-compile should have to have all sorts
of cross-platform header and library files installed. With this new
architecture, the burden rests more with the package authors, and
once the library files have been created, using this package should
be nearly as painless as using native Go packages.